finish raw version
