+++ src/components/Panorama.jsx
@@ -1,47 +1,21 @@
-import { useState, useMemo} from 'react';
+import { useMemo} from 'react';
 import { useLoader } from '@react-three/fiber';
 import * as THREE from 'three';
-import Box from './Box.jsx';
+import {OBJLoader} from "three/examples/jsm/loaders/OBJLoader.js";
 
 const backgrounds = [
     {
         id: 1,
         url: './sea.jpg',
     },
-    {
-        id: 2,
-        url: './pool.jpg',
-    },
-    {
-        id: 3,
-        url: './snow.jpg',
-    },
-    {
-        id: 4,
-        url: './skyscraper.jpg',
-    },
 ];
 
-export default function Panorama({ userImage }) {
-    const [selectImage, setSelectImage] = useState(userImage); 
-    console.log("Loading texture with URL:", userImage);
-    const [activeBackground, setActiveBackground] = useState(1);
-    // const { url } = backgrounds.find(({ id }) => id === activeBackground);
-    // const { url } = userImage ? userImage : backgrounds.find(({ id }) => id === activeBackground);
+export default function Panorama({ userImage, obj }) {
     const url = useMemo(() => {
-        return userImage || backgrounds.find(({ id }) => id === activeBackground).url;
-    }, [userImage, activeBackground]);
+        return userImage || backgrounds.find(({ id }) => id === 1).url;
+    }, [userImage]);
     const background = useLoader(THREE.TextureLoader, url);
-    const abc = selectImage? selectImage : './skyscraper.jpg'
-    const texture = useLoader(THREE.TextureLoader, abc);
-
-    // function getActiveBackground() {
-    //     var currentId = activeBackground + 1
-    //     if (currentId === 5) {
-    //         currentId = 1
-    //     }
-    //     return currentId
-    // }
+    const input_obj = useLoader(OBJLoader, obj)
 
     return (
         <group>
@@ -50,14 +24,7 @@ export default function Panorama({ userImage }) {
                 <meshBasicMaterial map={background} side={THREE.BackSide} />
             </mesh>
 
-            <group
-                onClick={(e) => {
-                    e.stopPropagation();
-                    // setActiveBackground(getActiveBackground());
-                }}
-            >
-                <Box />
-            </group>
+            <primitive object={input_obj}/>
         </group>
     );
 }
